[Development] Fix unbound `BUILDKITE` variable in `install-bazel.sh` (#35181)
The contribution docs instruct the user to run install-bazel.sh to install Bazel. However, this script errors with

install-bazel.sh: line 119: BUILDKITE: unbound variable
This PR fixes the error by providing a default value for the BUILDKITE variable when it's used in the script. The change ensures that if BUILDKITE is not set, it will be treated as an empty string, and the script will not throw an error.