[Core] Fix the recursion error when async actor has lots of deserialization. (#35494)
When an async actor is used, we always increase the max recursion limit before we post the function to the event loop because when we have lots of pending async tasks, Python thinks there's a recursion due to a large parallel callstacks (due to fiber which is used to implement async actor).

When running an async task, we have 3 steps.

run a deserialize function in the event loop,
ray/python/ray/_raylet.pyx

Line 866 in bfec451

 args = core_worker.run_async_func_in_event_loop( 
increase recursion limit,
ray/python/ray/_raylet.pyx

Line 831 in bfec451

 increase_recursion_limit() 
run a main function in the event loop
The problem here is that you increase the limit always "after" deserializing the object from the event loop. This means when the deserialization happens, the recursion limit is still low, and this can cause the exception.

When we have a lots of async tasks that has higher overhead of input deserialization, this can happen (because before we increase the recursion limit, we hit the max recursion error when deserializing the object), which is exactly what 1:1 async-actor calls with args async test does where the microbenchmark failed with the recursion error.

This PR fixes the issue by moving increase_recursion_limit inside the run_async_func_in_event_loop, so whenever we post a new async task, the recursion limit is always checked and increased.

I found the same issue when I developed a generator, and I verified this fixes the issue in this PR #35425 (comment).

I am not sure how to test this. @scv119 do you have the consistent repro that can run in unit tests? Or can you verify it using this branch as well?