[core] Don't drop rpc status in favor of reply status (#35530)
When an RPC returns, it returns a status, and a reply. Usually, the reply also contains a status.

Currently, if the status is not a GRPC error, we only propagate reply.status to the response callback. So, it is possible for the RPC status to be a non-GRPC error, while reply.status is OK. In this case, the RPC status is completely dropped, and the call is treated as successful.

Instead, we should:
1. Propagate the RPC status to the response callback if it is not ok.
2. If it is okay, propagate the reply status.

In this way, RPC status obscures reply status. Ideally we should come up with some way of setting status that doesn't allow one error to cover the other.