[serve] Move `num_queued_queries` logic out of `ReplicaSet` and into `Router` (#35436)
Centralizes all of the metrics definitions in one place, makes the `ReplicaSet` simpler. I plan to implement a new `ReplicaSet` for streaming and don't want to duplicate the metrics handling.

Also fixes a metrics bug: the `route` is dynamic and we are setting it as a tag for a global gauge. Removed that tag.