[Core] Fix test_no_worker_child_process_leaks (#35840)
Thanks @iycheng for debugging this!

In https://github.com/ray-project/ray/pull/33976 we don't always guarantee that children of worker processes are killed -- namely if the worker dies in an unexpected way, they are not (see https://github.com/ray-project/ray/issues/26118).

This unfortunately happens in the `test_no_worker_child_process_leaks` test case since the Ray cluster is torn down when the driver exits, which can lead to worker processes being forcefully terminated [the solution in this PR is to not initialize the Ray cluster in the driver script but start a separate head node with `.add_node`].

As a result of this forceful killing, the test becomes flaky (it passes if the workers are killed through the normal termination, and it doesn't pass if the workers are killed otherwise). The flakiness is not super big right now but becomes big with some other changes.

This should eventually be fixed with https://github.com/ray-project/ray/issues/26118.